requirements.txt
numpy
sounddevice
webrtcvad
faster-whisper
mcp
pydantic
scipy
pyttsx3
openai
google-generativeai
python-dotenv
edge-tts
duckduckgo-search
Pillow
gTTS
googlesearch-python
langgraph
langchain
langchain-google-genai
langchain-community
